+++ app/src/main/res/navigation/nav_graph.xml
@@ -72,6 +72,6 @@
         tools:layout="@layout/fragment_study_form" />
     <fragment
         android:id="@+id/dest_profile"
-        android:name="com.sesac.developer_study_platform.ui.ProfileFragment"
+        android:name="com.sesac.developer_study_platform.ui.profile.ProfileFragment"
         android:label="ProfileFragment" />
 </navigation>
